2024-12-16 - part 1 Add failing case #4
It's cheaper to get to (4,1) facing north via (4,2), but the overall
optimum path to E is through (3,1). This implies that graph nodes should
consist of a pair of the (x,y) and facing direction rather than only
considering the traversal direction in the cost function. That would
allow Djikstra's to evaluate the costs of each facing direction.